+++ packages/block-editor/src/components/letter-spacing-control/index.js
@@ -5,7 +5,6 @@ import {
 	__experimentalUnitControl as UnitControl,
 	__experimentalUseCustomUnits as useCustomUnits,
 } from '@wordpress/components';
-import deprecated from '@wordpress/deprecated';
 import { __ } from '@wordpress/i18n';
 
 /**
@@ -18,6 +17,9 @@ import { useSettings } from '../../components/use-settings';
  *
  * @param {Object}                  props                       Component props.
  * @param {boolean}                 props.__next40pxDefaultSize Start opting into the larger default height that will become the default size in a future version.
+ *
+ * @deprecated Default behavior since WordPress 7.1. Prop can be safely removed.
+ * @ignore
  * @param {string}                  props.value                 Currently selected letter-spacing.
  * @param {Function}                props.onChange              Handles change in letter-spacing selection.
  * @param {string|number|undefined} props.__unstableInputWidth  Input width to pass through to inner UnitControl. Should be a valid CSS value.
@@ -25,7 +27,7 @@ import { useSettings } from '../../components/use-settings';
  * @return {Element} Letter-spacing control.
  */
 export default function LetterSpacingControl( {
-	__next40pxDefaultSize = false,
+	__next40pxDefaultSize: _next40pxDefaultSize,
 	value,
 	onChange,
 	__unstableInputWidth = '60px',
@@ -37,23 +39,9 @@ export default function LetterSpacingControl( {
 		defaultValues: { px: 2, em: 0.2, rem: 0.2 },
 	} );
 
-	if (
-		! __next40pxDefaultSize &&
-		( otherProps.size === undefined || otherProps.size === 'default' )
-	) {
-		deprecated(
-			`36px default size for wp.blockEditor.__experimentalLetterSpacingControl`,
-			{
-				since: '6.8',
-				version: '7.1',
-				hint: 'Set the `__next40pxDefaultSize` prop to true to start opting into the new default size, which will become the default in a future version.',
-			}
-		);
-	}
-
 	return (
 		<UnitControl
-			__next40pxDefaultSize={ __next40pxDefaultSize }
+			__next40pxDefaultSize
 			__shouldNotWarnDeprecated36pxSize
 			{ ...otherProps }
 			label={ __( 'Letter spacing' ) }